requirements.txt
fastapi
uvicorn
torch
numpy
pycryptodome
python-dotenv
python-jose
passlib[argon2]
pytest
httpx
python-multipart
